Refactor the time helper to a shared function. Make the DNS healthcheck also report down based on the same time threshold
